move _get_units_hz_or_mhz to private _utils.py
