+++ src/nisarqa/processing/processing_utils.py
@@ -186,30 +186,4 @@ def clip_array(arr, percentile_range=(0.0, 100.0)):
     return np.clip(arr, a_min=vmin, a_max=vmax)
 
 
-def _get_units_hz_or_mhz(mhz: bool) -> tuple[str, str]:
-    """
-    Return the abbreviated and long units for Hz or MHz.
-
-    Parameters
-    ----------
-    mhz : bool
-        True for MHz units; False for Hz units.
-
-    Returns
-    -------
-    abbreviated_units : str
-        "MHz" if `mhz`, otherwise "Hz".
-    long_units : str
-        "megahertz" if `mhz`, otherwise "hertz".
-    """
-    if mhz:
-        abbreviated_units = "MHz"
-        long_units = "megahertz"
-    else:
-        abbreviated_units = "Hz"
-        long_units = "hertz"
-
-    return abbreviated_units, long_units
-
-
 __all__ = nisarqa.get_all(__name__, objects_to_skip)
